requirements.txt
alembic==0.7.6
bcrypt==2.0.0
cffi==1.1.2
click==4.1
flake8==2.4.1
-e git+https://github.com/mitsuhiko/flask.git@6e77cd709e16f7d1d654f67a5fc82e54bceae432#egg=Flask-dev
Flask-SQLAlchemy==2.0
functools32==3.2.3.post2
itsdangerous==0.24
Jinja2==2.7.3
jsonschema==2.5.1
Mako==1.0.1
MarkupSafe==0.23
mccabe==0.3.1
nose==1.3.6
pep8==1.5.7
psycopg2==2.6.1
pycparser==2.14
pyflakes==0.8.1
six==1.9.0
SQLAlchemy==1.0.4
structlog==15.2.0
Werkzeug==0.10.4
WTForms==2.0.2
